New version init commit
