+++ .travis.yml
@@ -1,7 +1,62 @@
+sudo: false
+
+dist: trusty
+
 language: php
 
 php:
-  - 5.6
-  - 7.0
+    - 5.6
+    - 7.0
+    - 7.1
+    - 7.2
+    - nightly
 
-script: true
+notifications:
+  email:
+    on_success: never
+    on_failure: change
+
+branches:
+  only:
+    - master
+    - develop
+
+cache:
+  apt: true
+  directories:
+    - vendor
+    - node_modules
+    - composer
+    - $HOME/.composer/cache
+
+matrix:
+  fast_finish: true
+  include:
+    - php: 7.2
+      env: WP_VERSION=latest WP_MULTISITE=0 SNIFF=1
+    - php: 7.1
+      env: WP_VERSION=latest WP_MULTISITE=0 SNIFF=1
+    - php: 7.0
+      env: WP_VERSION=latest WP_MULTISITE=1 SNIFF=1
+    - php: 5.6
+      env: WP_VERSION=latest WP_MULTISITE=1 SNIFF=1
+
+  allow_failures:
+    # Allow failures for unstable builds.
+    - php: nightly
+
+before_install:
+  # Speed up build time by disabling Xdebug.
+  # https://johnblackbourn.com/reducing-travis-ci-build-times-for-wordpress-projects/
+  # https://twitter.com/kelunik/status/954242454676475904
+  - phpenv config-rm xdebug.ini || echo 'No xdebug config.'
+  # Install PHP CodeSniffer.
+  - if [[ "$SNIFF" == "1" ]]; then composer self-update; fi
+  - if [[ "$SNIFF" == "1" ]]; then composer install; fi
+  - phpenv rehash
+
+script:
+  # Search for PHP syntax errors.
+  - find -L . -path ./vendor -prune -o -name '*.php' -print0 | xargs -0 -n 1 -P 4 php -l
+  # Run PHPCS.
+  - if [[ "$SNIFF" == "1" ]]; then vendor/bin/phpcs . --extensions=php --runtime-set ignore_warnings_on_exit 1; fi
